rename/remove mcp
