+++ dsc/src/main.rs
@@ -5,7 +5,7 @@ use args::{Args, SubCommand};
 use clap::{CommandFactory, Parser};
 use clap_complete::generate;
 use dsc_lib::progress::ProgressFormat;
-use mcp::start_mcp_server;
+use server::start_server;
 use rust_i18n::{i18n, t};
 use std::{io, process::exit};
 use sysinfo::{Process, RefreshKind, System, get_current_pid, ProcessRefreshKind};
@@ -19,9 +19,9 @@ use crossterm::event;
 use std::env;
 
 pub mod args;
-pub mod mcp;
 pub mod resolve;
 pub mod resource_command;
+pub mod server;
 pub mod subcommand;
 pub mod tablewriter;
 pub mod util;
@@ -86,9 +86,9 @@ fn main() {
             subcommand::function(&subcommand);
         },
         SubCommand::Server => {
-            if let Err(err) = start_mcp_server() {
+            if let Err(err) = start_server() {
                 error!("{}", t!("main.failedToStartServer", error = err));
-                exit(util::EXIT_MCP_FAILED);
+                exit(util::EXIT_SERVER_FAILED);
             }
             exit(util::EXIT_SUCCESS);
         }
