+++ dsc/src/server/list_dsc_resources.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-use crate::mcp::mcp_server::McpServer;
+use crate::server::mcp_server::McpServer;
 use dsc_lib::{
     DscManager, discovery::{
         command_discovery::ImportedManifest::Resource,
