+++ dsc/src/server/mcp_server.rs
